+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Snyk Changelog
 
+## [2.4.23]
+
+### Changed
+- Provide an additional info to the “Organization” setting
+
 ## [2.4.22]
 
 ### Fixed
@@ -49,7 +54,7 @@
 - Fix Container: invalid token shows error and does not redirect to Auth panel
 - Fix Container: should handle case if no images in project found
 - Fix Container: node still showing last results even if disabled
-- improved Snyk Container image parsing in K8S files
+- Improved Snyk Container image parsing in K8S files
 
 ## [2.4.17]